docs(resourcemanager, core): incorrect return value in `compareTo`
The docs of several `CompareTo` APIs says the return value is 0, 1, and -1. But actually for some use cases, the return value could be other numbers. We've found under some environment, our test cases will fail because the return value is 10, -10, 8, etc.
The test cases have been fixed. This commit will fix the docs of those API.