[Event Hubs] Support for Batch Processing
The focus of these changes is to expose the ability to create a checkpoint
with the `EventProcessorClient` as a protected member so that it can be
used when overriding `OnProcessingEventBatch` to work with batches rather
than individual events in the handler.

Previously, creating a checkpoint was bound to the event arguments passed
to the handler and required an awkward and complicated pattern to use.  By
making this protected, the `EventProcessorClient` can be extended in a manner
similar to `EventProcessor<T>` and allows working with batches without the
need to write a custom storage implementation.