Select phone numbers based on test agent
This avoids the need to make code changes when the test matrix or the phone numbers pool is modified.

This approach requires that the test agent identifier is explicitly defined as an environment variable in the job matrix in order to have a consistent way to get that information at runtime across different framework and OS versions.