[Event Hubs] Buffered Producer Await Handlers (#27173)
* [Event Hubs] Buffered Producer Await Handlers

The focus of these changes is to shift the invocation of event handlers
for the `EventHubBufferedProducerClient` from a non-deterministic background
approach to one that is deterministic and in consistent order.

Feedback from beta customers has indicated that the "random" nature of handler
invocations made it difficult to "checkpoint" with the source of events to
mark them as complete.  Handlers were potentially invoked in a different order
than the batches were sent, and the handler was unable to pause publishing for
a partition to accommodate completing a "checkpoint" operation.

Originally, the goal for background invocation was to ensure that handlers
which were long-running or performed a heavy workload could not impact
publishing throughput, but this also created additional pressure on the
thread pool due to the unconstrained use of `Task.Run` to ensure that handlers
written with synchronous code did not execute synchronously.