Update package version selection strategy
The verison selection process works as following:

Find at least one day old Azure.Core version so that other
packages have a chance to have build between the time core
and smoke test pipeline run.

Find other alpha packages that match exactly the same day
and if a package doesn't exist that matches the day then
we exclude it from the smoke testing for that day.

This will definitely lead to various package combinations but
they should all agree on version at least which will help with
the versioning conflicts.