[Event Hubs] Protected Idempotent Surface (#27914)
The focus of these changes is to remove the Experimental shims for exposing
the idempotent retries feature on the `EventHubConsumerClient` and shift
to a protected and non-browsable API on the client itself.

The API of the Experimental package was left unchanged for now to allow
partners the opportunity to change over to the new approach.  Once that
changeover is confirmed, the Experimental package will be retired.