[Event Hubs] Cancellation Token Exception Guards (#28113)
The focus of these changes is to guard against callbacks defined using
`CancellationToken.Register` in handlers which throw.  These exceptions
bubble up from developer code through the linked tokens and can cause
processor infrastructure to fail when the processor is stopping.

These guards will intercept the exceptions and log a warning before
ignoring them so that the shutdown process is able to complete normally.