smoke test pipeline hook (#25673)
* smoke test pipeline hook

* Package versioning

* Adding TimeOut

* Continue on error

* Formatting

* Added warnings and dynamic azure core min version

* Adding in Get-Package-Version functionality for release (non-Daily)

* [Smoke Tests] Restructure for Conflict Validation

The focus of these changes is to restructure and update the smoke tests
to validate that multiple Azure SDK packages can be used in the same application
without encountering conflicts for dependency versions.

This validation is focused only on ensuring that the assemblies do not
conflict and makes no attempt to validate scenarios from any specific
library.  Verifying library functionality is out of scope, and assumed to
be the responsibility of the library's test suite.

* pipeline test

* add install dotnet in smoke test

* smoke test pipeline hook

* Package versioning

* Adding TimeOut

* Continue on error

* Formatting

* Added warnings and dynamic azure core min version

* Adding in Get-Package-Version functionality for release (non-Daily)

* [Smoke Tests] Restructure for Conflict Validation

The focus of these changes is to restructure and update the smoke tests
to validate that multiple Azure SDK packages can be used in the same application
without encountering conflicts for dependency versions.

This validation is focused only on ensuring that the assemblies do not
conflict and makes no attempt to validate scenarios from any specific
library.  Verifying library functionality is out of scope, and assumed to
be the responsibility of the library's test suite.

* pipeline test

* add install dotnet in smoke test

* Test ignoring NU1605

* Exclude some references on unsupported platforms

* Fix constructor failure for TokenCache

Adding an additional corner-case guard for types that have a parameterless constructor but fail an internal state check.

* Additional edge cases for type/assembly loading.

* Additional "Not Found" corner cases

Handling additional "not found" cases for platform-specific assemblies during load or type activation.

* .NET Framework corner cases

Additional error handling for .NET Framework corner cases around attempting to load unsupported assembly types, such as UWP assemblies.

* .NET Framework activation guards for COM types

Additional guards for COM-based type activation scenarios for the .NET Framework target.

* .NET Framework fix for COM Error detection

* Additional context for ReflectionTypeLoadException

* Test run without recursive assembly discovery

* Removing recursive assembly discovery

I'm not convinced that the recursive assembly discovery is providing value; assemblies needed to reflect over types are being dynamically loaded on inspection.

* Fixing comment to agree with the code.

* Fixing logic error

* Update package version selection strategy

The verison selection process works as following:

Find at least one day old Azure.Core version so that other
packages have a chance to have build between the time core
and smoke test pipeline run.

Find other alpha packages that match exactly the same day
and if a package doesn't exist that matches the day then
we exclude it from the smoke testing for that day.

This will definitely lead to various package combinations but
they should all agree on version at least which will help with
the versioning conflicts.

* Add VSO formatting to logging

* Fixing log formatting

Error and Debug log formats were reversed.

* Client constructor now uses LogInformation

* Matrix cleanup

* Update common/SmokeTests/SmokeTest/SmokeTest.csproj

Co-authored-by: Heath Stewart <heaths@microsoft.com>

* Decrease output noise

* Adding in min matrix

Co-authored-by: Jesse Squire <jesse.squire@gmail.com>
Co-authored-by: Wes Haggard <Wes.Haggard@microsoft.com>
Co-authored-by: Jesse Squire <jsquire@microsoft.com>
Co-authored-by: Heath Stewart <heaths@microsoft.com>