[Event Hubs] Event Batch Enhancements (#28055)
* [Event Hubs] Event Batch Enhancements

The focus of these changes is to avoid paying the cost of serialization
from `EventData` instances to AMQP messages multiple times, where possible.

Previously, events were serialized to AMQP format when `TryAdd` was called
for the `EventDataBatch` in order to measure the event size, and then the
`EventData` instance was cloned and held in the batch.  When calling
`SendAsync`, the events held by the batch were again serialized to AMQP
format for each attempt at publishing; when retries were needed, the full
serialization pass was made against all batch events.

This approach was used for two reasons.  First, the AMQP transport tracks
delivery information on the serialized message directly and prevents
messages from being sent twice; in error scenarios, attempts to retry
sending were failing because of the transport behavior.  Additionally,
when idempotent retries are enabled, the metadata for sequencing events
must be applied as part of the send operation, preventing eager serialization
and caching.

For mainline scenarios, the changes included herein work around the
transport limitation by allowing previously serialized messages to be
packaged in a new AMQP batch envelope for each attempt to publish.  The
transport tracks delivery on the envelope, which is discarded in the
event of retries and is rebuilt without needing to perform a full
serialization pass over the batch events.  This removes the need to clone
the source event in `TryAdd` and allows the batch to hold the serialized
AMQP message, eliminating the need to perform serialization for batch
events when `SendAsync` is called.

When idempotent retries are enabled, many of the benefits are negated;
the idempotent metadata written to an event cannot be calculated eagerly
and must be determined at the time `SendAsync` is called.  Because of this,
the previous behavior of cloning `EventData` instances in `TryAdd` and
holding those instances in the batch is retained.   Additionally, a full
serialization pass over the batch attempts is neeed when `SendAsync` is
called.  Some efficiency is gained in retry scenarios, as the previously
serialized messages can be reused with a new AMQP envelope for each
publishing attempt rather than performing a full serialization pass against
the source events as had been previously done.

* Refactoring to use a factory method for creating the correct AmqpTransportEventBatch subtype

* Fixing minor formatting error

* WIP | Tests pass

* Refactoring idempotent implementation

This allows AMQP messages to be cached in idempotent scenarios, with
responsibility for applying sequencing shifted to the batch rather than
being purview of the producer.

* Adding idempotent property size calculation back into the transport batch.

* Removing the ability to request a collection of EventData from the batch in the model factory to conform behavior to the actual type.