[Text Analytics] Legacy API and Abstractions
The focus of these changes is add the generated surface for the the
legacy Text Analytics REST API in a stand-alone project that is compatible
with the AutoRest generator and its conventions.  The types for the legacy API
are generated under a unique namespace and will be included in the Text Analytics
client library via source linking.  This is intended to allow the legacy API
to be managed and regenerated in a standard way while avoiding conflicts
with the Cognitive Language API that will be used for service operations on
new version going forward.

Also included are the stubs for the `ServiceClient` abstractions, forming
the basis of the contract by which the `TextAnalyticsClient` will be able to
target different REST services through a common interface.