Event Hubs: Buffered Producer Improvements
The focus of these changes is to enhance the buffered producer with some
minor improvements based on stress and performance test findings.

- When the service is throttling, logic has been added to allow for retrying
with an additional back-off until the operation completes or has been cancelled.
Because callers do not directly influence when publishing takes place, they
should not be exposed to failures for throttling; doing so would require
that they take extreme steps, such as manually publishing the throttled
events or recreating their producer to reset buffers if the ordering of
events needed to be restored.

- When events are not available in the buffer, an expanding back-off is
now applied to the publishing loop.  This is intended to avoid wasting
resources in a spin-wait scenario during periods of low activity.  In the
current implementation, consecutive empty buffer checks will result in a
delay pattern that looks like:

    - 25 ms
    - 35 ms
    - 48 ms
    - 68 ms
    - 96 ms
    - 134 ms
    - 188 ms
    - 263 ms
    - 368 ms
    - 516 ms
    - 723 ms
    - 1 second

Once the back-off reaches one second, it will continue to use that value
until an event is available in the buffer.

- The approach to setting buffered producer-specific retry defaults is now
more robust and will be applied when options are passed rather than only in
the case when they were not.  Customization of retry values will continue
to be appropriately respected.