Fixing CodeGen Capitalization
The focus of these changes is to fix a typo in the `CustomEntitiesResultDocumentsItem` stub which was causing "item" to be generated as lower-case.  This differs from the capitalization of the file committed to the repository and was causing issues with case-sensitive file systems in CI when the "Analyze" step was run during API diff checks.