[Event Hubs Extensions] Bump E2E Test Timeout (#28932)
Bumping the timeout used for the end-to-end live tests by 15 seconds to
reduce flakiness.  Currently, the same value is used for the timeout that
Event Hubs uses for the operation timeout.  When the network is slow or
overloaded, this prevents the client from attempting a retry.

The new value allows a small window above the operation timeout which
allows the client to perform one retry in timeout scenarios, which often
is successful.