Merge CLU 1.0.0 release branch to main
Because CLU 1.1 beta was meant to ship before CLU 1.0 GA due to features
announced at MSBuild, it was originally in main while I created a
release branch for CLU 1.0 GA to stabilize it; however, a decision was
made to ship CLU as a DPG though it wasn't complete. As issues were
found and the 1.0 GA date approached, I needed to switch gears to 1.0 GA
while DPG was still taking updates. This is better done in main.

I pushed main to a release/clu-2022-05-15-preview branch and did a
non-merge - git checkout release/clu-2022-05-01 -- sdk/cognitivelanguage
- to main because the history wasn't worth all the merge conflicts and
  manual updates (mostly deletes) requires to actually merge. Hence a
  bit of history here, in case its ever needed.