Merge CLU 1.0.0 release branch to main (#29358)
Because CLU 1.1 beta was meant to ship before CLU 1.0 GA due to features announced at MSBuild, it was originally in main while I created a release branch for CLU 1.0 GA to stabilize it; however, a decision was made to ship CLU as a DPG though it wasn't complete. As issues were found and the 1.0 GA date approached, I needed to switch gears to 1.0 GA while DPG was still taking updates. This is better done in main.

This PR is a literal copy of the release/clu-2022-05-01 branch I created above. Because we prefer a linear history, dealing with all the merge conflicts that would abound wasn't worth it. So I created this branch / PR by:

1. Deleting everything in _sdk/cognitivelanguage_.
2. Ran `git checkout release/clu-2022-05-01 -- sdk/cognitivelanguage` to copy all files from that branch's HEAD.
3. Ran `eng/scripts/CodeChecks.ps1 cognitivelanguage` to make sure generated code was up to date, despite having just recently taking the same autorest.csharp version into my previous release/clu-2022-05-01 branch.

Since history of my commits - which should have helpful descriptions - will be lost, I wanted to include a bit of the history here in case it's ever required for troublshooting.