[Event Hubs] Flaky Test Tweaks
The focus of these changes is to tweak some tests that have been flaky during
nightly runs.  The timeout for Functions end-to-end tests has been bumped
slightly to allow additional time for an event to be made available for reading
in the partition, since this is not governed by an SLA on the service-side.
The buffered producer idle test been focused on its core scenario, removing
one of the synchronization gates to help streamline and keep forward progress.