[Event Hubs] Processor Improvements
The focus of these changes is to fix an error with startup validation where
an invalid consumer group or container was not detected.  Additionally, a
warning was added which is surfaced to the event processor error handler
signaling that an unhandled exception propagated from the event processing
handler causing partition processing to fault and restart.

Riding along are fixes to a couple of broken links that failed validation
and troubleshooting guide improvements.