[Event Hubs] Update Scale Monitor
The focus of these changes is to update the scale monitor for the Event
Hubs Function extension to correct some left over T1 patterns that are
no longer applicable.  For example, it is valid to have null checkpoints
for a partition, but not possible to have a checkpoint with null offset.

Additionally, correcting the assumption that the beginning sequence number
for a partition is always 0.  This may not be true if the partition has been
previously used and events have aged out, such as for very low volume or
DR/fail-over scenarios.