[Event Hubs] Data body copy for EventDataBatch
The focus of these changes is to avoid an unintentional loophole allowing
an event batch to be mutated by mutating the data body of an `EventData`
instance that has been added to it.  When an `EventData` is serialized into
the interim `AmqpMessage` format used by the transport library, a projection
over the existing `ReadOnlyMemory<byte>` buffer has been used to avoid
allocations.

To combat this, the batch will now make a defensive copy of the data body
of an `EventData` when accepted into the batch, using a shared buffer that
has been rented.  When the batch is disposed, the buffer is returned.

**Additional context:**

Though the buffer is typed as `ReadOnlyMemory<byte>`, since the it was
passed to the EventData constructor, and the scope of the "call" is the
lifetime of the associated instance.  We explicitly intend that the `EventData`
instance is no longer relevant to the batch after `TryAdd` completes.
Because of this, it should not matter what a caller does with the body buffer
after, the batch should be fully immutable.