[Event Hubs] Tune processor ownership docs
The focus of these changes is to add guidance for choosing the number of
processor instances to the configuration sample and reduce the frequency
that the "owns too many partitions" warning is logged.  It will now occur
only when the owned partition count changes, rather than on each load
balancing cycle.