[Event Hubs] Live test fixes
The focus of these changes is to make tests more deterministic and
resilient to the service returning duplicates, which should be expected
behavior.

Previously, tests use the count of events read as the condition to signal
that reading should cease and validation begin.  When duplicates were
present, this caused some expected events to not appear in the final set
being validated.

The approach for validating reads from enqueue time were also adjusted to
account for clock skew and a starting point that may not be precise, where
the service begins reading a bit earlier than expected, which would also
cause additional events to be read and throw off validation.