Switch to script over pwsh to enable command logging
If we use the script block instead of pwsh the DevOPs task
will output the script command line in the logs which is very
helpful when trying to reproduce a test failure locally.

As part of this also need to switch to double quotes as single
quotes don't work in the script block correctly.