Refactor the build method of ServiceManager to remove internal dependency
In ServiceManagerStore.cs, refactor the build method of ServiceManager to remove internal dependency. The original implementation builds the ServiceManager from a ServiceCollection directly, and inject IConfigureOptions<ServiceManagerOptions> and IOptionsChangeTokenSource<ServiceManagerOptions> to support hot reload. Without internal depdencies, we have to use the ServiceManagerBuilder to support hot-reload. The only way of ServiceManagerBuilder to support hot-reload is to inject a IConfiguration object which provides a change token. However, the SDK and function extensions have different ways to parse configuration, for example, SDK lacks the functionality to parse identity-based connection from configuration. Therefore, we have to pass the actual configuration action via ServiceManagerBuilder.WithOptions, and injects an IConfiguration object to provide change token. To avoid errors thrown by SDK when it parses the configuration, we wraps the real IConfiguration object to make it looks like empty configuration but provides the true change token.