Make sure backup/restore tests are non-parallel
Get frequent conflicts when running live tests; though, I'm not sure how
much this will help because the RecordingTestBase itself is marked as
[NonParallelizable], which is an inheritted attribute. Won't hurt,
though. Could just be a service issue because, sometimes, an LRO may not
be completely done when it says it is and we're stressing the service
pretty hard (and rather impractically).