[Event Hubs Bindings] Parameter Precedence Fix (#33813)
The focus of these changes is to fix a bug in the client factory where
behavior differed when creating clients when a connection string and
event hub name were both configured.

The producer and consumer gave precedence to the connection string, as
documented, allowing both to be specified without error.  The event
processor gave precedence to the event hub value, causing potential errors
to surface when it disagreed with the connection string.