[Event Hubs] Allow partition key use in async collector
The focus of these changes is to allow for partition keys to be used when
binding to the async collector to publish events.  Previously, use was
restricted to the `IAsyncCollector<T>` interface, which only accepted an
`EventData` instance leaving no way to specify a partition key.  As a result,
all events were assigned partitions via round-robin.

These changes expose the `EventHubAsyncCollector` and allow binding to it,
exposing an `AddAsync` overload specific to Event Hubs which accepts a
partition key.  This allows events to be logically grouped together, ensuring
they are assigned to the same partition.