[Smoke Tests] Ignore Known Incompatible Packages
The focus of these changes is to define a set that contains the list of
packages which should not be included as references to the smoke tests.
This is necessary because some compatibility packages do not target
netstandard2.0, and the package source is unable to filter upstream to
restrict to the appropriate packages for the running target.