[Text Analytics] Remove options parameter from `DynamicClassify` and `DynamicClassifyAsync` methods (#33915)
The `DynamicClassify` and `DynamicClassifyAsync` methods of the `TextAnalyticsClient` class should not receive a `DynamicClassifyOptions` parameter. This is to align with the other convenience methods that receive a single document as a string parameter (e.g., `RecognizeEntities`, `ExtractKeyPhrases`, etc.)

The `ClassificationType` property is also being "moved out" of the `DynamicClassifyOptions` class and turned into a standalone parameter that can be passed into all the dynamic classification methods. This means that customers can still choose the classification type (multi vs. single) when calling `DynamicClassify` or `DynamicClassifyAsync` without the options parameter. Additionally, we no longer need the `DynamicClassifyOptions` sub-class, since `ClassificationType` was the only property that it added on top of the `TextAnalyticsRequestOptions` base class.