Remove DECODE_PRECENT variable after agent update (#34028)
After the changes to the KeyVault task at https://github.com/microsoft/azure-pipelines-tasks/pull/17566 we need to remove this opt-out otherwise we now end up with values that are not decoded which need to be. This became an issue for any values with a % that were set with the VSO setvariable syntax.