fix(build): cadl build error when repo url does not end with `.git`
`[.git]` is a char set which means the url should ends with either `.`, or `g`, or `i`, or `t`. So if the url doesn't end with `.git` then there is no match which results in error.

The fix is to remove `[.git]` char set. We don't need `(.git)*` either, because `\S+` will suppress it.