[Event Hubs] Function Test Fix
The focus of these changes is to fix an assumption in the Functions extension
package that the same Storage suffix is valid in all clouds.  The implementation
supports the Azure public cloud only, while the test was asserting that
the scenario was valid for all clouds.  Going forward, the test will be ignored
when the current execution environment is not a supported cloud.

Also riding along is a tweak to a processor test to extract additional
error details for failure investigation.