Fix #33474 by changing preprocessor statement.
As discussed in the bug, LoginWithPromptAsync is only available on .NET Framework <4.6 right now because the preprocessor statement targets net452 explicitly. I believe that to be a mistake, so I've changed the preprocessor statement to target any version of the Framework.

I also took the liberty of deleting a little bit of commented code.