[Event Hubs] Tweak processor startup
The focus of these changes is to tweak the approach used to validate
connection permissions when the processor is first starting up.  Previously,
it would attempt a read using the same explicit consumer used when running.
That would potentially cause the processor working on that partition to
temporarily be pushed off the partition.  Though it would recover on the next
load balancing cycle, this would cause a delay in processing.