[Event Hubs Trigger] MinBatch WaitTime Adjustments (#35369)
* [Event Hubs Trigger] MinBatch WaitTime Adjustments

The focus of these changes is to guard against the scenario where
partition ownership changes while the trigger is waiting to build a batch
of the requested size.  Because the new owner has no insight into how long
the previous owner has already waited, the wait time will restart and obey
the configured `MaxWaitTime`.  This could result in a batch for the
partition not being dispatched in a timely manner.   In the worst case,
where ownership of the partition changes multiple times, this could lead
to a batch for the partition appearing to be severely delayed or to never
be dispatched to the Function.

To guard against this, the trigger will now consider the time that the
most recently read checkpoint was created and prefer dispatching a partial
batch when the partition reads events for the first time and the checkpoint
is older than the configured `MaxWaitTime` or when no checkpoint
was found.