[EngSys] Remove Outdated Stress Framework (#36064)
* [EngSys] Remove Outdated Stress Framework

The focus of these changes is to delete the outdated stress framework that
should no longer be used and remove references to it.  The stress projects
removed were confirmed to have been empty from the example template and
did not contain any actual tests.  It has also been confirmed that this change
does not impact code generation.