Potential wrong use of assertion (#34574)
We encountered an error while using the SDK where GetAll would always throw. After looking through the code we saw that where an instance of this class is created (TrafficManagerProfileResource), the ctor with the 3rd parameter is never called. I'm guessing that's from the AutoRest codegen.

Maybe adding AssertNotNull rather than AssertNull would fix the problem in the codegen?

How we made it work was to use reflection to get _profileData and set it before calling GetAll()