[Management] Remove legacy testcommon path
The focus of these changes is to remove the legacy `/testcommon` path from
the repository.  This was originally used to bootstrap development of the
T2 management libraries and contained initial test infrastructure and cases.
It has since become outdated and is no longer used.  The `Azure.Graph.Rbac`
test package used mimics a T2 client library and is potentially confusing.