[Event Hubs] Restore stress project references
The focus of these changes is to restore the direct reference to the Event
Hubs core project from the stress project.  To avoid versioning conflicts,
a directory props file has been added that forces all stress project
dependencies to use project references.  This will ensure that the references
to the Event Hubs core package from stress and the processor are consistent.