Remove System.Net.Http package dependency
Fixes https://github.com/Azure/azure-sdk-for-net/issues/36627

Some of the libraries in this repository package reference System.Net.Http which then gets added as a dependency transitively in the consuming app. A package reference to System.Net.Http is undesireable as it brings in the entire netstandard1.x dependency graph and slows down restore. Additionally, potential future CVEs (and alerts surfaced by Component Governance) wouldn't impact libraries in this repository as the dependency would be avoided entirely.

The API exposed by System.Net.Http is already inbox in modern TFMs like .NETCoreApp, netstandard2.x and net461 and above and doesn't need to be brought in via a package reference.