[Messaging] Mocking sample annotation
The focus of these changes is to annotate the mocking samples for Event Hubs
and Service Bus with a call-out to the general Azure SDK unit testing and
mocking article.

While the article is focused around HTTP services and
does not replace these dedicated samples, the additional context and examples
may be helpful to developers.