[AppConfig] Hide internal constructor from ConfigurationSetting
The focus of these changes is to hide constructor of `ConfigurationSetting`
that allows the service owned ETag to be set.  This was intended to be an
internal constructor and was inadvertently made public.

In addition, the ability to create `FeatureFlagConfigurationSetting` and
`SecretRefrenceConfigurationSetting` instances from the model factory has
been added.