[Event Hubs] Log improvements + Amqp version bump
The focus of these changes is to improve logs for the event processor
types and AMQP consumer.  The most notable change is the addition of the
first and last sequence number of the batch when reading from Event Hubs
and dispatching events to be processed.

The version of the AMQP library has been bumped to the latest stable (2.6.3),
which includes a fix for timer duration calculations during link creation
and a small handful of efficiency improvements.