[Functions] Event Hubs: Adjust default batch size
The focus of these changes is to increase the default for the maximum batch
size from 10 to 100.

This setting represents the maximum number of events from Event Hubs that
the function can receive when it's invoked. The decision to change the
default was based on developer feedback, testing, and a desire to match
the defaults used by the Azure SDK for Event Hubs.  This change will be
beneficial to the average scenario by helping to improve performance
as well as lower costs due to fewer function executions.