[AppConfig] Preserve unknown setting JSON members (#38450)
* [AppConfig] Preserve unknown setting JSON members

The focus of these changes is to fix a bug where JSON members in the
value for Feature Flags and Secret References which are not defined in
the JSON schema were discarded.  The schema itself allows for arbitrary
JSON members, the Azure portal supports them in its advanced view, and
the service documentation also refers to them.

This change preserves unknown JSON attributes while allowing the well-known
properties of the setting to override their associated JSON members.

Constructor overloads that accept an ETag were also added to Feature Flags
and Secret References.  This aligns with ConfigurationSetting and allows use
of the GetConfigurationSettingsAsync overload that accepts `onlyIfUnchanged`,
which was previously not accessible for specialized settings.