[Azure.Core] DRAFT - Response Disposal for Anne
The purpose of these changes is to illustrate a flawed work-around for
the "protocol methods return a disposed response" problem.  I do not
advocate this for the fix, rather an illustration of the problem that
works in the success case.

This is not intended to be merged, it exists only for discussion.