[AppConfig] Allow unknown condition members
The focus of these changes is to allow for unknown members to be
present in the `conditions` property of a feature flag configuration
setting.  Previously, these were ignored as the setting assumed that
`conditions` should be treated as a well-known member.  This change moves
the well-known member to `conditions/client_filters` and preserves unknown
children of `conditions`.