[Event Hubs] Fix ETW arg type inference
The focus of these changes is to fix an issue with the inferred type for
an ETW event argument, which was causing the event to be interpreted as
a Unicode string rather than a double.  This caused the message text to
display incorrectly and the value was represented as the wrong type in the
structured arguments array.