[Event Hubs] Error handling for disabled entities (#39715)
* [Event Hubs] Error handling for disabled entities

The focus of these changes is to add dedicated handling when an Event Hub
resource has been set to "disabled" status.  This is now treated as a terminal
error with a "ResourceNotFound" status, as it requires manual intervention.
Previously, this would result in a general communication error that was
transient and triggered implicit retries.