Pass token credential to share client builder
When using the Azure client builder `AddFileServiceClient` extension
method to configure a `ShareServiceClient` using the file service URI,
the `TokenCredential` instance was not being passed to the
`ShareServiceClient` constructor. Consequently, instances of
`ShareClient` created from it were unable to authenticate because they
did not have access to the credential configured in the client builder.

This commit ensures that `ShareServiceClient` is constructed with the
credential so that its `ShareClient` instances can use the token for
authentication.