Address pull request comments
This commit adds a new extension method to the Azure Files Shares client
factory builder that allows for the creation of a `ShareServiceClient`
by URI that uses the configured token credential. This revision avoids
a breaking change that would have been caused by changing the generic
constraint of the existing extention method.