[Event Hubs] Processor fixes
The focus of these changes is to address a race condition where calling
`IsRunning` concurrently with starting/stopping the processor could cause a
semaphore to be double-released.  As part of this fix, the approach used to
detect when a semaphore should be released has been changed to consider the
semaphore count rather than using a sentinal variable.

Also fixed was an issue where the processor would leak a quota exceeded
exception during startup validation if there were more than 4 other processors
starting concurrently.  The partition chosen for validation is now randomized and
a quota exceeded response is now considered a successful validation.